Consolidate evaluator planning tests
