+++ tests/table/test_evaluator_planning.py
@@ -22,7 +22,7 @@ from collections.abc import Callable
 import pytest
 
 import pyiceberg.table as table_module
-from pyiceberg.expressions import BooleanExpression, GreaterThan
+from pyiceberg.expressions import BooleanExpression, EqualTo, GreaterThan
 from pyiceberg.manifest import DataFile, FileFormat
 from pyiceberg.schema import Schema
 from pyiceberg.table import ManifestGroupPlanner, Table
@@ -63,3 +63,32 @@ def test_partition_evaluator_prepares_once_per_spec(table_v2: Table, monkeypatch
     assert not partition_evaluator(_data_file(1, 1))
     assert partition_evaluator(_data_file(2, 10))
     assert evaluator_calls == [[1, 10]]
+
+
+def test_metrics_evaluator_prepares_once_per_plan(table_v2: Table, monkeypatch: pytest.MonkeyPatch) -> None:
+    class CountingMetricsEvaluator:
+        def __init__(
+            self,
+            schema: Schema,
+            expr: BooleanExpression,
+            case_sensitive: bool = True,
+            include_empty_files: bool = False,
+        ) -> None:
+            self.calls: list[DataFile] = []
+            instances.append(self)
+
+        def eval(self, data_file: DataFile) -> bool:
+            self.calls.append(data_file)
+            return True
+
+    instances: list[CountingMetricsEvaluator] = []
+    monkeypatch.setattr(table_module, "_InclusiveMetricsEvaluator", CountingMetricsEvaluator)
+    planner = ManifestGroupPlanner(table_metadata=table_v2.metadata, io=table_v2.io, row_filter=EqualTo("x", 10))
+    first_file = _data_file(1, 1)
+    second_file = _data_file(2, 2)
+
+    metrics_evaluator = planner._build_metrics_evaluator()
+    assert len(instances) == 1
+    assert metrics_evaluator(first_file)
+    assert metrics_evaluator(second_file)
+    assert instances[0].calls == [first_file, second_file]
